+++ NEWS.md
@@ -1,5 +1,8 @@
 # tidyr 0.8.2.9000
 
+* `crossing()` (and hence `expand()`) now works with list-columns 
+  (#446, @SamanthaToet).
+
 * `crossing()` preserves factor levels (#410).
 
 # tidyr 0.8.2